Steemauto Privacy Policy
Privacy policy is a very important document on any website that details a company's views on information collected from visitors and how it intends to secure that information. It also helps to instill more confidence in the prospective site visitors/users and give users an idea of what to expect and what not to expect on the site.